refactor(analytics): 例外 import を canonical owner へ移行する (#3954)
